Update README.md
1)Wine quality prediction:
This assignment is using information like alcohol, acidity, volatility to predict the quality of wine. In this project. Linear Regression is used to predict the quality of wine 
2) Gold Price Prediction:
This assignment is predicting the price of gold. Here the Random Forest technique is used as it works well with large continuous datasets. 
3. Credit Card Fraud Detection:
This project is  predicting whether a given transaction is fraudulent or fair. Here Logistics Regression, model is used